author changes to data loss patch
